file_forge/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.32.0
pydantic>=2.10.1
python-multipart>=0.0.21
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.36
alembic>=1.14.0
cryptography>=44.0.0
aiofiles>=24.1.0
requests>=2.32.3
python-dotenv>=1.0.1
httpx>=0.28.1
aiohttp>=3.11.11
pytest>=8.3.3
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
redis>=5.2.1
celery>=5.4.0
pillow>=11.1.0
pytesseract>=0.3.13
onnxruntime>=1.20.1
openpyxl>=3.1.5
python-docx>=2.0.0
pdfplumber>=0.11.4

# AI/ML Dependencies
transformers>=4.48.0
torch>=2.5.1
sentence-transformers>=3.1.1
faiss-cpu>=1.8.0
openai>=1.55.0
langchain>=0.3.9
langchain-openai>=0.2.8
langchain-community>=0.3.7
scikit-learn>=1.6.1
joblib>=1.4.2
numpy>=1.26.4
timm>=1.0.11
accelerate>=0.34.2
bitsandbytes>=0.44.2

# Specialized File Processing Dependencies
# CAD/Engineering
EzDXF>=0.18.1
octave>=0.18.2

# E-books & Documents
beautifulsoup4>=4.12.3
lxml>=5.3.0
mobi>=0.4.1

# Medical/Compliance
pydicom>=2.4.4
pylibjpeg>=1.4.0
python-gdcm>=3.0.24

# Geospatial
geopandas>=1.0.1
shapely>=2.0.6
pyproj>=3.7.0
fiona>=1.10.1

# Disk Images
pycdlib>=1.15.2
pypff>=2024.5.30

# Email
extract-msg>=0.46.5
mailbox>=0.5.0
chm>=0.40.0

# Security fixes for transitive dependencies
ecdsa>=0.19.0
urllib3>=2.2.2
anyio>=4.2.0
sentry-sdk>=2.0.0
python-magic>=0.4.27
elasticsearch>=8.0.0
